requirements.txt
web3
eth-account
cryptography
pyyaml
requests
colorama
tabulate
python-dotenv
